Refactors tokenizer to emit a Token double-linked-list
This refactors the tokenizer, simplifying both tokenizer and parser code slightly (fewer utility functions) and returns tokens as part of the AST's `loc` information. Tokens also have gained a few new properties: `line` for the 1-indexed line they appear on and `prev`/`next` to navigate the dll of Tokens. Finally, this introduces three new Token types: `<SOF>` for the empty token at beginning of the file stream, `,` for ignored comma tokens, and `Comment` for ignored comment tokens. These tokens are *included* in the Token list allowing for easier navigation of the parsed document in GraphQL tools and as a basis for parsing comment descriptions in the schema language.